streaming/spark-streaming/examples/producer/requirements.txt
boto3
kafka-python
